wip(db): Update database import in progress

+++ CS209A-Proj-Backend/crawler/src/main/kotlin/io/github/octcarp/sustech/cs209a/proj/crawler/model/BriefUser.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BriefUser(
-    val userId: Int? = null,
+    val userId: Long? = null,
 )
